gunyah: vm: tests: Correct size in unmap_memory_single
This test tries some simple unmap operations and ensures they fail. One
of the unmap operations tries to unmap 10MB offset within the
guestmemfd. The guestmemfd was not made large enough (11MB) to create
the guest memory region and thus fails before we can make the unmap
call. Correct the overall size of the guestmemfd.

Signed-off-by: Elliot Berman <elliot.berman@oss.qualcomm.com>